+++ crates/oas3/src/spec/media_type_examples.rs
@@ -81,9 +81,8 @@ impl MediaTypeExamples {
     }
 }
 
-#[cfg(test)]
+#[cfg(all(test, feature = "yaml-spec"))]
 mod tests {
-    #[cfg(feature = "yaml_spec")]
     use serde_json::json;
 
     use super::*;
@@ -99,7 +98,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "yaml_spec")]
     fn deserialize() {
         assert_eq!(
             serde_yaml::from_str::<MediaTypeExamples>(indoc::indoc! {"
@@ -121,7 +119,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "yaml_spec")]
     fn serialize() {
         let mt_examples = MediaTypeExamples::default();
         assert_eq!(
@@ -143,7 +140,6 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "yaml_spec")]
     fn single_example() {
         let spec = serde_yaml::from_str::<Spec>(indoc::indoc! {"
 openapi: 3.1.0
@@ -171,7 +167,6 @@ paths: {}
     }
 
     #[test]
-    #[cfg(feature = "yaml_spec")]
     fn resolve_references() {
         let spec = serde_yaml::from_str::<Spec>(indoc::indoc! {"
 openapi: 3.1.0