+++ crates/oas3/src/spec/schema.rs
@@ -593,7 +593,7 @@ where
     T::deserialize(de).map(Some)
 }
 
-#[cfg(all(test, feature = "yaml_spec"))]
+#[cfg(all(test, feature = "yaml-spec"))]
 mod tests {
     use super::*;