+++ crates/roast/examples/complex-validation.rs
@@ -1,9 +1,11 @@
 #![allow(dead_code, unused_variables)]
 
+use std::fs;
+
 use http::Method;
 use serde_json::json;
 
-fn main() {
+fn main() -> eyre::Result<()> {
     let _ = dotenvy::dotenv();
     pretty_env_logger::init();
 
@@ -29,15 +31,16 @@ fn main() {
         "page": 1
     });
 
-    let spec = oas3::from_path("./data/oas-samples/allof.yml").expect("api spec parse error");
+    let yaml = fs::read_to_string("./data/oas-samples/pet-store.yml")?;
+    let spec = oas3::from_yaml(yaml).expect("api spec parse error");
     let op = spec.operation(&Method::GET, "/").unwrap();
-    let schema = &op.responses(&spec)["200"].content["application/json"]
-        .schema(&spec)
-        .unwrap();
+    let schema = &op.responses(&spec)["200"].content["application/json"].schema(&spec)?;
 
     // let v = ValidatorTree::from_schema(&schema, &spec).unwrap();
     // v.validate(valid).unwrap();
     // v.validate(invalid1).unwrap();
     // v.validate(invalid2).unwrap();
     // v.validate(invalid3).unwrap();
+
+    Ok(())
 }